temp. removed async. runs and detects

+++ ttsmanager.py
@@ -1,10 +1,8 @@
-import asyncio
-import json
+import json #for serializing objects
 import pyttsx3 #for the TTS engine
+import os #for finding files
 
 class TTSManager:
-    file_lock = asyncio.Lock()
-    tts_lock = asyncio.Lock()
     catchphrases = { #TODO load from file instead
         "Tiffany" : "borbnation will flourish.",
         "Patrick" : "slayer of bugnation.",
@@ -20,20 +18,20 @@ class TTSManager:
 
     def __init__(self):  
         self.engine = pyttsx3.init() #intialize TTS engine
-        with open("catchphrases.json", "rt") as f:
-            catchphrases = json.load(f)
+        if os.path.exists(self.params.CATCHPHRASES_DB_PATH):
+            with open("catchphrases.json", "rt") as f:
+                catchphrases = json.load(f)
     
     def __del__(self): #TODO
-
         #write catch phrases to db file
-        async with file_lock:
+        if os.path.exists(self.params.CATCHPHRASES_DB_PATH):
             with open(self.params.CATCHPHRASES_DB_PATH, "wt") as f:
                     json.dump(self.catchphrases, f)
 
-    async def request_speak(self, speech): #TODO
-        async with tts_lock:
-            self.engine.say(speech)
-            self.engine.runAndWait()
+    def request_speak(self, speech): #TODO
+    #async with tts_lock:
+        self.engine.say(speech)
+        self.engine.runAndWait()
 
     async def add_catchphrase(self, name, catchphrase):
-            self.catchphrases[name] = catchphrase
+        self.catchphrases[name] = catchphrase
